api/pyproject.toml
[tool.poetry]
name = "meals"
version = "1.0.0"
description = ""
authors = ["Peter Havens <peter.havens@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.1"
Django = "^4.0.1"
psycopg2 = "^2.9.3"
django-environ = "^0.9.0"
djangorestframework = "^3.13.1"
django-extensions = "^3.1.5"
celery = {extras = ["redis"], version = "^5.2.3"}

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
watchdog = {extras = ["watchmedo"], version = "^2.1.6"}
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"
pytest-clarity = "^1.0.1"
pytest-freezegun = "^0.4.2"
pytest-mock = "^3.6.1"
factory-boy = "^3.2.1"
pytest-celery = "^0.0.0"
flake8-logging-format = "^0.6.0"
pytest-profiling = "^1.7.0"
bandit = "^1.7.4"
flake8-bugbear = "^22.3.23"
flake8-bandit = "^3.0.0"
ipython = "^8.3.0"
ipdb = "^0.13.9"
Werkzeug = "^2.1.2"
django-types = "^0.15.0"
djangorestframework-types = "^0.8.0"
celery-types = "^0.13.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "config.settings"
